Implement basic resource management system functionality
This solution has serializer and deserializer for assets. Assets are stored in a bundle file and afterwards loaded into the RAM in runtime.
For now this is not efficient and there are a lot of gaps to be solved later on.
There's no guarantee that this code works, after google tests adding it will be tested